#85 updated coments on AppointmentPopup and AppointmentPicker.kt
